refactor: Move tool definitions to a separate module
- Moved the `tool` decorator and all tool functions from `kodeagent.py` to a new `tools.py` module to improve modularity.
- Updated `__init__.py` to export the tools from the new module, maintaining the public API.
- Updated unit tests to reflect the new module structure.
- Updated `README.md` to point to the new `tools.py` module for examples.
- Added a module-level docstring to `tools.py`.
