+++ src/kodeagent/__init__.py
@@ -1,16 +1,7 @@
 """KodeAgent: An intelligent code agent"""
 
 from .kodeagent import (
-    tool,
     llm_vision_support,
-    calculator,
-    search_web,
-    download_file,
-    search_arxiv,
-    extract_file_contents_as_markdown,
-    search_wikipedia,
-    get_youtube_transcript,
-    get_audio_transcript,
     Agent,
     ReActAgent,
     CodeActAgent,
@@ -26,6 +17,17 @@ from .kodeagent import (
     CodeRunner,
     AgentResponse,
     print_response
+)
+from .tools import (
+    tool,
+    calculator,
+    search_web,
+    download_file,
+    search_arxiv,
+    extract_file_contents_as_markdown,
+    search_wikipedia,
+    get_youtube_transcript,
+    get_audio_transcript,
 )
 from .kutils import (
     is_it_url,
